+++ libsrc/ffdec_lib/test/com/jpexs/decompiler/flash/as3decompile/ActionScript3ClassicDecompileTest.java
@@ -1984,6 +1984,15 @@ public class ActionScript3ClassicDecompileTest extends ActionScript3DecompileTes
     }
 
     @Test
+    public void testTernarOperator2() {
+        decompileMethod("classic", "testTernarOperator2", "var b:Boolean = true;\r\n"
+                + "var i:int = 1;\r\n"
+                + "var j:int = b ? i : i + 1;\r\n"
+                + "var k:int = Boolean(i) ? j : j + 1;\r\n",
+                 false);
+    }
+
+    @Test
     public void testTry() {
         decompileMethod("classic", "testTry", "var i:int = 0;\r\n"
                 + "i = 7;\r\n"