+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Fixed
 - [#2424] DefineEditText handling of letterSpacing, font size on incorrect values
+- [#2391] Double not operator in ternar operator expression
 
 ## [22.0.2] - 2025-01-17
 ### Added
@@ -3691,6 +3692,7 @@ Major version of SWF to XML export changed to 2.
 [alpha 8]: https://github.com/jindrapetrik/jpexs-decompiler/compare/alpha7...alpha8
 [alpha 7]: https://github.com/jindrapetrik/jpexs-decompiler/releases/tag/alpha7
 [#2424]: https://www.free-decompiler.com/flash/issues/2424
+[#2391]: https://www.free-decompiler.com/flash/issues/2391
 [#2375]: https://www.free-decompiler.com/flash/issues/2375
 [#2374]: https://www.free-decompiler.com/flash/issues/2374
 [#2389]: https://www.free-decompiler.com/flash/issues/2389